chore(SDK): drop 252 and promote 261 to stable
