+++ base/BUILD
@@ -519,7 +519,6 @@ java_library(
         "//intellij_platform_sdk:plugin_api_for_tests",
         "//intellij_platform_sdk:test_libs",
         "//proto:proto_deps",
-        "//sdkcompat:testFramework",
         "//shared:artifact",
         "//shared:vcs",
         "//testing:lib",
