+++ base/tests/unittests/com/google/idea/blaze/base/util/VersionCheckerTest.java
@@ -126,16 +126,6 @@ public class VersionCheckerTest {
       throw new UnsupportedOperationException();
     }
 
-    // @Override #api252 remove this function
-    public @Nullable String getProductUrl() {
-      throw new UnsupportedOperationException();
-    }
-
-    // @Override #api252 remove this function
-    public @Nullable String getJetBrainsTvUrl() {
-      throw new UnsupportedOperationException();
-    }
-
     @Override
     public boolean hasHelp() {
       throw new UnsupportedOperationException();
