+++ clwb/src/com/google/idea/blaze/clwb/radler/RadTestContextProvider.kt
@@ -27,7 +27,7 @@ import com.google.idea.blaze.base.run.producers.TestContext
 import com.google.idea.blaze.base.run.producers.TestContextProvider
 import com.google.idea.blaze.base.util.pluginProjectScope
 import com.google.idea.blaze.cpp.CppBlazeRules.RuleTypes
-import com.google.idea.sdkcompat.radler.RadTestPsiElement
+import com.intellij.clion.radler.testing.RadTestPsiElement
 import com.intellij.execution.actions.ConfigurationContext
 import com.intellij.util.asSafely
 import com.intellij.util.io.await
