--- intellij_platform_sdk/BUILD.clion252
@@ -1,136 +1,0 @@
-# Copyright 2024 The Bazel Authors. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-# Description:
-#
-# Plugin source jars for CLion, accessed remotely.
-
-load("@rules_java//java:defs.bzl", "java_import")
-
-package(default_visibility = ["//visibility:public"])
-
-java_import(
-    name = "sdk",
-    jars = glob([
-            "lib/*.jar",
-            "lib/modules/*.jar",
-        ],
-        exclude = [
-            "lib/testFramework.jar",
-        ],
-    ),
-    tags = ["intellij-provided-by-sdk"],
-    deps = ["@error_prone_annotations//jar"],
-)
-
-java_import(
-    name = "test_framework",
-    jars = ["lib/testFramework.jar"],
-)
-
-java_import(
-    name = "guava",
-    jars = glob([
-        "lib/lib.jar",
-        "lib/lib-client.jar",
-    ]),
-)
-
-java_import(
-    name = "clion_plugins",
-    jars = glob([
-        "plugins/c/lib/*.jar",
-        "plugins/c/lib/modules/*.jar",
-        "plugins/cidr-base/lib/*.jar",
-        "plugins/cidr-base/lib/modules/*.jar",
-        "plugins/cidr-lang/lib/*.jar",
-        "plugins/cidr-lang/lib/modules/*.jar",
-        "plugins/clion-cmake/lib/*.jar",
-        "plugins/clion-cmake/lib/modules/*.jar",
-        "plugins/clion-test-google/lib/*.jar",
-        "plugins/clion-test-google/lib/modules/*.jar",
-        "plugins/clion-test-catch/lib/*.jar",
-        "plugins/clion-test-catch/lib/modules/*.jar",
-        "plugins/clion-test-boost/lib/*.jar",
-        "plugins/clion-test-boost/lib/modules/*.jar",
-        "plugins/clion-radler/lib/*.jar",
-        "plugins/clion-radler/lib/modules/*.jar",
-        "plugins/clion-rd-components/lib/*.jar",
-        "plugins/nativeDebug-plugin/lib/*.jar",
-        "plugins/clion/lib/*.jar",
-        "plugins/clion/lib/modules/*.jar",
-        "plugins/platform-images/lib/*.jar", # for native debug plugin
-        "plugins/performanceTesting/lib/*.jar",
-        "plugins/performanceTesting/lib/modules/*.jar",
-        "plugins/textmate/lib/*.jar",
-        "plugins/textmate/lib/modules/*.jar",
-    ]),
-)
-
-filegroup(
-    name = "clion_tools",
-    srcs = glob(["bin/**/*"]),
-)
-
-java_import(
-    name = "javascript",
-    jars = glob(["plugins/javascript-plugin/lib/*.jar"]),
-)
-
-java_import(
-    name = "css",
-    jars = glob(["plugins/css-impl/lib/*.jar"]),
-)
-
-java_import(
-    name = "tslint",
-    jars = glob(["plugins/tslint/lib/*.jar"]),
-)
-
-java_import(
-    name = "terminal",
-    jars = glob(["plugins/terminal/lib/terminal.jar"]),
-)
-
-java_import(
-    name = "python",
-    jars = glob(["plugins/python-ce/lib/*.jar"]),
-)
-
-filegroup(
-    name = "python_helpers",
-    srcs = glob(["plugins/python-ce/helpers/**/*"]),
-)
-
-java_import(
-    name = "bundled_plugins",
-    jars = glob([
-       "plugins/clion/lib/*.jar",
-       "plugins/clion/lib/modules/*.jar",
-    ]),
-    tags = ["intellij-provided-by-sdk"],
-)
-
-filegroup(
-    name = "application_info_json",
-    srcs = ["product-info.json"],
-)
-
-java_import(
-    name = "textmate",
-    jars = glob([
-        "plugins/textmate/lib/*.jar",
-        "plugins/textmate/lib/modules/*.jar",
-    ]),
-)
