--- sdkcompat/v253/com/google/idea/sdkcompat/clion/debug/CidrDebuggerPathManagerAdapter.kt
@@ -1,28 +1,0 @@
-/*
- * Copyright 2026 The Bazel Authors. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.google.idea.sdkcompat.clion.debug
-
-import com.jetbrains.cidr.execution.debugger.CidrDebuggerPathManager
-import java.io.File
-
-// #api252
-object CidrDebuggerPathManagerAdapter {
-
-  fun getBundledGDBBinary(): File {
-      return CidrDebuggerPathManager.bundledGDBBinary
-  }
-}
