--- sdkcompat/v261/com/google/idea/sdkcompat/ide/AbstractWizardAdapter.java
@@ -1,32 +1,0 @@
-/*
- * Copyright 2026 The Bazel Authors. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.google.idea.sdkcompat.ide;
-
-import com.intellij.ide.wizard.AbstractWizard;
-import com.intellij.ide.wizard.Step;
-import com.intellij.openapi.project.Project;
-import com.intellij.openapi.util.NlsContexts.DialogTitle;
-import org.jetbrains.annotations.NonNls;
-import org.jetbrains.annotations.Nullable;
-
-// #api252
-public abstract class AbstractWizardAdapter<T extends Step> extends AbstractWizard<T> {
-
-  public AbstractWizardAdapter(@DialogTitle String title, @Nullable Project project) {
-    super(title, project);
-  }
-}
