--- sdkcompat/v261/com/google/idea/sdkcompat/radler/module.kt
@@ -1,22 +1,0 @@
-/*
- * Copyright 2026 The Bazel Authors. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.google.idea.sdkcompat.radler
-
-typealias RadSymbolHost = com.intellij.clion.radler.core.protocol.RadSymbolsHost
-typealias RadTestPsiElement = com.intellij.clion.radler.testing.RadTestPsiElement
-
-typealias RadProjectFallbackConfigDisabler = com.intellij.rider.cpp.core.projectModel.RadProjectFallbackConfigDisabler
